requirements.txt
# Core dependencies
transformers>=4.36.0
torch>=2.0.0
torchaudio>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.24.0
scipy>=1.11.0

# Development tools
python-dotenv>=1.0.0

# AWS S3 access
boto3>=1.26.0